# openapi_dump.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TraderCopilot Backend",
    "version": "2.0.1"
  },
  "paths": {
...
    "/news/": {
      "get": {
        "tags": [
          "News"
        ],
        "summary": "Get News",
        "operationId": "get_news_news__get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Items limit",
              "default": 20,
              "title": "Limit"
            },
            "description": "Items limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/advisor/chat": {
      "post": {
        "tags": [
          "Advisor",
          "advisor"
        ],
        "summary": "Advisor Chat",
        "description": "Endpoint for interactive chat with the Advisor.\nEnforces Entitlements:\n1. Plan Access (can_use_advisor)\n2. Daily Quota (advisor_chat)\n3. Token Access (if context token provided)",
        "operationId": "advisor_chat_advisor_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/advisor/profile": {
      "get": {
        "tags": [
          "Advisor",
          "advisor"
        ],
        "summary": "Get Advisor Profile",
        "description": "Get the user's Copilot profile. Creates a default one if it doesn't exist.",
        "operationId": "get_advisor_profile_advisor_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CopilotProfileResp"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Advisor",
          "advisor"
        ],
        "summary": "Update Advisor Profile",
        "description": "Update the user's Copilot profile.",
        "operationId": "update_advisor_profile_advisor_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CopilotProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CopilotProfileResp"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/advisor/": {
      "post": {
        "tags": [
          "Advisor",
          "advisor"
        ],
        "summary": "Analyze Advisor",
        "description": "Analiza una posici\u00f3n abierta y sugiere alternativas.\nVersi\u00f3n local determinista (sin LLM).\nEnforces Entitlements.",
        "operationId": "analyze_advisor_advisor__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvisorReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdvisorReq": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token",
            "description": "Token base: eth|btc|sol|xau"
          },
          "direction": {
            "type": "string",
            "enum": [
              "long",
              "short"
            ],
            "title": "Direction",
            "description": "Direcci\u00f3n de la posici\u00f3n"
          },
          "entry": {
            "type": "number",
            "title": "Entry",
            "description": "Precio de entrada de la posici\u00f3n"
          },
          "tp": {
            "type": "number",
            "title": "Tp",
            "description": "Take profit previsto"
          },
          "sl": {
            "type": "number",
            "title": "Sl",
            "description": "Stop loss previsto"
          },
          "size_quote": {
            "type": "number",
            "title": "Size Quote",
            "description": "Tama\u00f1o de la posici\u00f3n en moneda de cotizaci\u00f3n (ej. USDT)"
          }
        },
        "type": "object",
        "required": [
          "token",
          "direction",
          "entry",
          "tp",
          "sl",
          "size_quote"
        ],
        "title": "AdvisorReq"
      },
...
      "ChatContext": {
        "properties": {
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          },
          "timeframe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeframe"
          },
          "signal_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Data"
          }
        },
        "type": "object",
        "title": "ChatContext"
      },
      "ChatMessage": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage"
      },
      "ChatRequest": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "type": "array",
            "title": "Messages"
          },
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatContext"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "ChatRequest"
      },
      "CopilotProfileResp": {
        "properties": {
          "trader_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trader Style",
            "default": "BALANCED"
          },
          "risk_tolerance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Tolerance",
            "default": "MEDIUM"
          },
          "time_horizon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Horizon",
            "default": "INTRADAY"
          },
          "custom_instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Instructions"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id"
        ],
        "title": "CopilotProfileResp"
      },
      "CopilotProfileUpdate": {
        "properties": {
          "trader_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trader Style",
            "default": "BALANCED"
          },
          "risk_tolerance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Tolerance",
            "default": "MEDIUM"
          },
          "time_horizon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Horizon",
            "default": "INTRADAY"
          },
          "custom_instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Instructions"
          }
        },
        "type": "object",
        "title": "CopilotProfileUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/token"
          }
        }
      }
    }
  }
}